WIP of database modifications
Added fields to User table and updated methods accordingly.
Added getUsers(String selection) for all user related queries.
Modified addUser to include all parameters. An unique user id is
calculated using the device id and the number of users that share that
id. The deviceId has to be your ownId for it to be unique.